feat: currentUser 함수로 로그인 유무와 회원정보 확인
